Add support to create a new document
This commit adds a very simple and basic interface to create a new
metanorma document, it currently supports csd document but very soon
we will add all of the other supported document.

The adds a temporary new executable `exe/metanorma-new` and this
will allow us to create a new document, and the way we can do it
now is to use the following command.

```sh
exe/metanorma-new new -t csd -d standard my-csd-document
```

Fair warning, this is still in very early stage to support our use
case, but we will soon clean up this and make it as friendly as
possible.