Adopt thor gem for the metanorma cli
The current version of metanorma cli is more likely a single
ruby file, and we are doing all the option parsing and others
manually. This also makes it bit harder to extend, and also
forces us to re-write some common behavior by ourself.

This commit changes this, and adopt the most commonly used
library `thor` for this purpose, it also migrate all existing
behavior to this new one

Since `metanorma` is being used in production, and this new
version is still in early stage, so instead of replacing it
directly, we are doing a soft release. If you wanna give it
a shot then you can do so by setting up an environment var
`METANORMA_DEV_MODE` to `true`.