Remove doctype from generated document
Currently the generated document contains a folder structure by
the doctype and then have it's item nested in it, but we actually
don't want it but to keep a flatten hierarchy.

This commit add that functionality, and instead of nesting the
files under doctype folder, it will now directly add those to
the root, and keep the other structure as it is.