Fix all migration related issues
During the migration of the new version, we've added some default
value to some of the options and that's what was breaking some of
the existing functionality.

We were also passing the options keys as string, but previously
we were passing it as symbols, and that was also creating some
issue in this new version.

This commit fixes all of these issues, and it tries to bring it
as close as possible toward the existing version.