Add support to use git template for document
Currently, we have some basic standard templates included with the
CLI and user can only use any of those to create a new document, but
in reality we want the user to be able to provide custom template and
use it to create new document.

This commit add that functionality, so user can provide custom git
template repository as cli argument and metanorma will automatically
download and use that template for document creation.

This also stores those template in the cli, so in the future if
it's desirable then we can also use this cache version.