Add custom message for missing file
Currently, if there are no proper file provided with the command
then we always show the help documents, or invoke version related
command. But, we actually do want to show some specific message

This commit customizes it and if there are more attributes than
the default one and missing a file then it will show the file
missing error.