Gracefully fail for invalid commands
Since, we have invoked our custom command as root command, so
every single command is going through our main compile command
and this shouting out loud in the console.

This commit changes this behavior, so if user pass an invalid
command or that's not a document name then it will show the
error and suggest them to use the `metanorma help` instead.

Fixes - 52