Fix require library `(-r)` issue
In document compilation we missed the `require` options, and it
was not behaving as expected and introduced issue like 49. This
commit fixes this so this should work as expected from now on.

This also updates the existing `overwrite` options, it adds`-y`
as an aliases and also changes it to be a flag.

Fixes 49