Add interface to list supported extensions
Currently, we do not have any easier way for user to know the
supported extensions for a type. Which might actually confuse.

This commit adds an interface, which allows a user to retrieve
the list of supported extensions for a type. For example: if
you want to check the supported extensions for `iso` then you
can use the following.

```
metanorma list-extensions iso
```

Closes: 76