List supported extensions for all type
The CLI offers `list-extensions type` interface to display all
supported extensions for a specific type, but we want to extend
this feature.

This commit improve this interface, so if the user specify a
single type then it will display supported extensions for that
specific type, but if they don't then it will display all of
the supported extensions. Usages:

```sh
metanorma list-extensions
```

Closes #78