Remove old codes from the executable
There was some major migration in the CLI lately, but we kept
some of the existing codes as there was an way to switch back
to that one using an environment variable.

But since the new version has been published for a while now
and it is stable, so this commit removes those existing code
form the library.

This also removes the `bundler/setup` from the executable

Closes #47