Fix inconsistency for version interface
There are some inconsistency in the `version` interface, one of
command shows version for default backend, but other doesn't.

This commit changes this, so if the user doesn't provide any
type, it will display versions for all supported dependencies.
But user can also retrieve a specific version by adding it to
the type option.

Related: #98