Remove actual network call from test suite
Context:

Currently, the fontist related tests were performing actual network
call for each request. Which is not ideal, and it is expected from
the library to return paths of a fonts, not running the full cycle
of downloading a fonts. That is already being tests in fontist.

Solution:

This commit removes this network calls, and instead of running the
whole cycle it mocks the expected behavior, and only test what the
metanorma cli is epxected to do. Since most of those fonts should
not be distributable by third party so it also usages a fake font
for the tests.

Fixes 164