Add an interface to generate a mini site
Site generation is another task that we want to handle though the
Metanorma CLI. The way it should work is: we would pass a source
directory and we want metanorma to find each of the files and run
the complier to compile those to an expected output.

Once the complication is done then we also want it to invoke the
Relaton library to build the collection and finally convert the
collection to a base html page.

We also wanted this interface to support a configuration file,
so each site generation will also have more control over the
internal of site generation. This is exactly what this commit
tried to achieve, so now a user can generate a mini site using:

```sh
metanorma site generate SOURCE_PATH -o OUTPUT_PATH -c manifest.yml
```