Add support for wildcards in file selection
The earlier version of mini site configuration file only supports
a list with exact filenames, but we want this to support wildcard
so we don't have to specify every single file.

This commit adds the support for this wild card pattern, and it
also ensures we handle the duplicates gracefully.