Fix include related compilation issues
We are running the compilation and collection building inside from the
expected output directory, this is good for isolations during the site
generation phase but it was not correctly resolving the file path for
the compiler.

This commit fixes this issue by moving out the compile step from that
isolation and use an absolute path for compiled output.

One note though, it turns out the compile interface is not behaving
as expected when you are not in the source directory, and run compile
interface for a file that includes include directive. This could be
an issue with the compile interface or how AsciiDoc handle include.

Fixes - 196