Add support to respect  with default manifest
During porting the `site` interface to build scripts, I realized
all of the samples, and templates repos are using a default config
file called `metanorma.yml`. To comply with this all actions need
to pass this file to the site generator which creates some sort of
dependency, especially if a repo doesn't have a config file.

This commit changes this, so when no config file is provided then
it will also try to find a manifest in the current directory, and
if there is one then it will use that one for site generation.