Cleanup the metanorma functional spec files
Overtime the `spec/metanorma_spec` grew a lot and it becoming hard
to maintain and also understand what's going on that. So, this commit
tries to clean it up a bit and also simplify we were doing lots of
stuff there.

As a side node, it made me thing if we actually need to have this
test file with this repository, given that this are already being
tested in the metanorma library. These are bit repetitive and form
the CLI prospective it only delegates some messages, and those are
already being tested in the acceptance tests, but I would leave it
it like this for now I guess.