Add support for wildcard document selection
Currently, the `metanorma compile` interface allow us to pass
one single file at a time but in reality we might want to compile
multiple files, and based on the current setup user have to write
some sort of script to lop through those files.

This commit adds that support as built in with the `compile` command,
so if a user passes a wildcard character with the filename then it'll
try to find all the matches and compile them one by one.

Fixes 204