Add default source path for site interface
The current site interface expect us to pass a specific path, but
this might feel redundant especially when we are already in the
present working directory and we still have to pass `.` just to
meet that requirement.

This commit changes this, and makes the `source_path` optional,
so now user doesn't need to provide `.` for present working
directory.

Fixes #208