Add support for custom template in site
The current site interface is using a default template we have in
the relaton library, this works great but we want to give more control
to the actual user so they can use their custom styles or template

This commit adds the support for the style and template options, and
once the user pass this options then we pass it down to relaton to
handle it accordingly.

Normal usages wise the options are `template_dir`, and `stylesheet`
but user can also use the `metanorma.yml` file to more stable config.

```ruby
metanorma:
  ...
  template:
    path: "./path-to-your-template"
    stylesheet: "./path-to-your-stylesheet-file"
```

Doc: https://github.com/relaton/relaton-cli#relaton-xml2html