Explicitly activate all flavored gems
There has been a long pending issue with the performance for the
cli, especially if we have multiple version installed. In that case
lots of gem still stays as unresolved, and that could cause some of
the performance issue.

This commit follows some suggestion to explicitly activate those
gems, as it should resolve the dependencies early. This should also
fix the private gem loading issue as we have seen on another repo

Related: https://github.com/metanorma/metanorma-ribose/issues/83