Add simplified docker setup for development
With the frequent changes in the project it gets bit hard sometime to
setup everything in a local machine, and also test it out in multiple
ruby environment when we need to debug things.

This commit adds a simplified way to run the project in docker and it
also give us full control to switch between the version easily, so we
hope it will help us to debug things easily.

Note: We are explicitly including files for the gem, so this folder
should not be shipped to the actual gem, but there is some changes
then we can adjust those in the future.