Gracefully handle keyboard interruption
Let's say, we are running the `metanorma` command, and for whatever
reason if we want to exit the process (CTRL-C, and others) and don't
continue with the command then it blows up with error with in the
terminal with useless backtracing information.

This commit changes that, so now if we try to exit the process then
it should gracefully handle the signal, show a simple message.