Fix collection option related issue
We have recently added support to provide collection options either
as an command line arguments or options in the collection file. But
that is only working on the site interface, and the collection cli
still expect the `output-folder` as required option.

This commit changes that, so it's not required anymore and it will
try to read the option from the file, and if none of those source
has this value, then it will use the source as output directory.

Fixes #283